+++ lib/class/playlist.class.php
@@ -91,8 +91,11 @@ class Playlist extends playlist_object
     /**
      * get_playlists
      * Returns a list of playlists accessible by the user.
+     * @param boolean $incl_public
+     * @param int $user_id
+     * @return array
      */
-    public static function get_playlists($incl_public = true, $user_id = null)
+    public static function get_playlists($incl_public = true, $user_id = -1, $playlist_name = '')
     {
         if (!$user_id) {
             $user_id = Core::get_global('user')->id;
@@ -107,13 +110,21 @@ class Playlist extends playlist_object
 
         if ($incl_public) {
             if (count($params) > 0) {
-                $sql .= ' OR ';
+                $sql .= " OR `type` = 'public'";
             } else {
-                $sql .= ' WHERE ';
+                $sql .= " WHERE `type` = 'public'";
             }
-            $sql .= "`type` = 'public'";
+        }
+
+        if ($playlist_name !== '') {
+            if (count($params) > 0 || $incl_public) {
+                $sql .= " AND `name` = '" . $playlist_name . "'";
+            } else {
+                $sql .= " WHERE `name` = '" . $playlist_name . "'";
+            }
         }
         $sql .= ' ORDER BY `name`';
+        debug_event('playlist.class', 'get_playlists query: ' . $sql, 5);
 
         $db_results = Dba::read($sql, $params);
         $results    = array();
@@ -437,8 +448,8 @@ class Playlist extends playlist_object
          */
         $sql        = "SELECT `track` FROM `playlist_data` WHERE `playlist` = ? ORDER BY `track` DESC LIMIT 1";
         $db_results = Dba::read($sql, array($this->id));
-        $data       = Dba::fetch_assoc($db_results);
-        $base_track = $data['track'] ?: 0;
+        $track_data = Dba::fetch_assoc($db_results);
+        $base_track = $track_data['track'] ?: 0;
         debug_event('playlist.class', 'Adding Media; Track number: ' . $base_track, 5);
 
         $count = 0;